[EventDispatcher] Added EventTrait
This allows to easily make any custom class fit for EventDispatcher
