+++ src/Symfony/Component/EventDispatcher/Debug/TraceableEventDispatcher.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\EventDispatcher\Debug;
 
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Symfony\Component\EventDispatcher\EventInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\EventDispatcher\Event;
 use Symfony\Component\Stopwatch\Stopwatch;
@@ -118,7 +119,7 @@ class TraceableEventDispatcher implements TraceableEventDispatcherInterface
     /**
      * {@inheritdoc}
      */
-    public function dispatch($eventName, Event $event = null)
+    public function dispatch($eventName, EventInterface $event = null)
     {
         if (null === $event) {
             $event = new Event();
@@ -220,20 +221,20 @@ class TraceableEventDispatcher implements TraceableEventDispatcherInterface
     /**
      * Called before dispatching the event.
      *
-     * @param string $eventName The event name
-     * @param Event  $event     The event
+     * @param string         $eventName The event name
+     * @param EventInterface $event     The event
      */
-    protected function preDispatch($eventName, Event $event)
+    protected function preDispatch($eventName, EventInterface $event)
     {
     }
 
     /**
      * Called after dispatching the event.
      *
-     * @param string $eventName The event name
-     * @param Event  $event     The event
+     * @param string         $eventName The event name
+     * @param EventInterface $event     The event
      */
-    protected function postDispatch($eventName, Event $event)
+    protected function postDispatch($eventName, EventInterface $event)
     {
     }
