+++ src/Symfony/Component/EventDispatcher/ImmutableEventDispatcher.php
@@ -38,7 +38,7 @@ class ImmutableEventDispatcher implements EventDispatcherInterface
     /**
      * {@inheritdoc}
      */
-    public function dispatch($eventName, Event $event = null)
+    public function dispatch($eventName, EventInterface $event = null)
     {
         return $this->dispatcher->dispatch($eventName, $event);
     }
